Fixed issue with Rename Library Added more verbose error when there are library import errors Improved use of test libraries in Unit Tests New UnitTests to check for library correct behaviour (helped a lot to debug library rename issue)

+++ Web Applications/TeamMentor.CoreLib/TM_AppCode/XmlDatabase/TM_Xml_Database.Library.GuidanceItem.cs
@@ -126,32 +126,35 @@ namespace TeamMentor.CoreLib
     {				
         public static TM_Xml_Database           xmlDB_Load_GuidanceItems(this TM_Xml_Database tmDatabase)
         {
-            var pathXmlLibraries = TM_Xml_Database.Current.Path_XmlLibraries;            
-            if (pathXmlLibraries.notNull() && pathXmlLibraries.notNull())
-                lock (pathXmlLibraries)
-                {
-                    if (tmDatabase.getCacheLocation().fileExists().isFalse())
+            if (tmDatabase.UsingFileStorage)
+            { 
+                var pathXmlLibraries = TM_Xml_Database.Current.Path_XmlLibraries;            
+                if (pathXmlLibraries.notNull() && pathXmlLibraries.notNull())
+                    lock (pathXmlLibraries)
                     {
-                        "[TM_Xml_Database] in xmlDB_Load_GuidanceItems, cache file didn't exist, so creating it".debug();
-                        var o2Timer = new O2Timer("loaded GuidanceItems from disk").start();
-                        //Load GuidanceItem from the disk				
-                        foreach (var guidanceExplorer in tmDatabase.xmlDB_GuidanceExplorers())
+                        if (tmDatabase.getCacheLocation().fileExists().isFalse())
                         {
-                            var libraryId = guidanceExplorer.library.name.guid();
-                            var pathToLibraryGuidanceItems = pathXmlLibraries.pathCombine(guidanceExplorer.library.caption);
-                            "libraryId: {0} : {1}".info(libraryId, pathToLibraryGuidanceItems);
-                            var filesToLoad = pathToLibraryGuidanceItems.files(true, "*.xml");
-                            tmDatabase.xmlDB_Load_GuidanceItemsV3(libraryId, filesToLoad);
+                            "[TM_Xml_Database] in xmlDB_Load_GuidanceItems, creating cache file".debug();
+                            var o2Timer = new O2Timer("loaded GuidanceItems from disk").start();
+                            //Load GuidanceItem from the disk				
+                            foreach (var guidanceExplorer in tmDatabase.xmlDB_GuidanceExplorers())
+                            {
+                                var libraryId = guidanceExplorer.library.name.guid();
+                                var pathToLibraryGuidanceItems = pathXmlLibraries.pathCombine(guidanceExplorer.library.caption);
+                                "libraryId: {0} : {1}".info(libraryId, pathToLibraryGuidanceItems);
+                                var filesToLoad = pathToLibraryGuidanceItems.files(true, "*.xml");
+                                tmDatabase.xmlDB_Load_GuidanceItemsV3(libraryId, filesToLoad);
+                            }
+
+                            //save it to the local cache file (reduces load time from 8s to 0.5s)
+                            tmDatabase.save_GuidanceItemsToCache();
+                            o2Timer.stop();
+
+                            tmDatabase.ensureFoldersAndViewsIdsAreUnique();
+                            tmDatabase.removeMissingGuidanceItemsIdsFromViews();
                         }
-
-                        //save it to the local cache file (reduces load time from 8s to 0.5s)
-                        tmDatabase.save_GuidanceItemsToCache();
-                        o2Timer.stop();
-
-                        tmDatabase.ensureFoldersAndViewsIdsAreUnique();
-                        tmDatabase.removeMissingGuidanceItemsIdsFromViews();
                     }
-                }
+            }
             return tmDatabase;
         }        
         public static List<TeamMentor_Article>  xmlDB_Load_GuidanceItemsV3(this TM_Xml_Database tmDatabase, Guid libraryId, List<string> guidanceItemsFullPaths)
